Move validation to module service layer
