+++ packages/core/core-flows/src/translation/steps/index.ts
@@ -5,4 +5,3 @@ export * from "./validate-translations"
 export * from "./create-translation-settings"
 export * from "./update-translation-settings"
 export * from "./delete-translation-settings"
-export * from "./validate-translation-settings"
