+++ packages/core/core-flows/src/translation/workflows/update-translations.ts
@@ -7,7 +7,6 @@ import {
 } from "@medusajs/framework/workflows-sdk"
 import { emitEventStep } from "../../common/steps/emit-event"
 import { updateTranslationsStep, UpdateTranslationsStepInput } from "../steps"
-import { validateTranslationsStep } from "../steps"
 import { TranslationWorkflowEvents } from "@medusajs/framework/utils"
 
 /**
@@ -22,13 +21,13 @@ export const updateTranslationsWorkflowId = "update-translations"
  *
  * You can use this workflow within your own customizations or custom workflows, allowing you
  * to update translations in your custom flows.
- * 
+ *
  * @since 2.12.3
  * @featureFlag translation
  *
  * @example
  * To update translations by their IDs:
- * 
+ *
  * ```ts
  * const { result } = await updateTranslationsWorkflow(container)
  * .run({
@@ -61,11 +60,6 @@ export const updateTranslationsWorkflow = createWorkflow(
   (
     input: WorkflowData<UpdateTranslationsWorkflowInput>
   ): WorkflowResponse<TranslationDTO[]> => {
-    const validateInput = transform(input, (input) => {
-      return "translations" in input ? input.translations : [input.update]
-    })
-    validateTranslationsStep(validateInput)
-
     const translations = updateTranslationsStep(input)
 
     const translationIdEvents = transform(
